scripts/requirements.txt
openpyxl>=3.1.2
xlsxwriter>=3.2.0
reportlab>=4.1.0
python-pptx>=0.6.23
Pillow>=10.2.0
Faker>=22.0.0
